+++ packages/twig-to-php-parser/__tests__/fixtures/larva.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+	parser: {
+		twigDir: `${__dirname}/src/patterns/`,
+		phpDir: `${__dirname}/template-parts/patterns/`,
+	}
+};