+++ packages/twig-to-php-parser/__tests__/twig-to-php-parser.test.js
@@ -7,10 +7,8 @@ const exec = require( 'child_process' ).exec;
 const fixture = path.join( __dirname, 'fixtures' );
 const expectedDir = `${fixture}/template-parts-expected/patterns/`;
 
-const config = {
-	twigDir: `${fixture}/src/patterns/`,
-	phpDir: `${fixture}/template-parts/patterns/`,
-};
+const getAppConfiguration = require( '../lib/getConfig' );
+const appConfiguration = getAppConfiguration( 'parser' );
 
 const patternShortPaths = [
 	'components/c-byline.php',
@@ -26,28 +24,28 @@ describe( 'twig to php parser', function() {
 
 	beforeEach( ( done ) => {
 
-		exec( 'mkdir ' + config.phpDir, ( err  ) => {
+		exec( 'mkdir ' + appConfiguration.phpDir, ( err  ) => {
 			if ( err ) {
 				console.error( err );
 			}
 		});
 
-		return twigToPhpParser( config )
+		return twigToPhpParser( appConfiguration )
 		.catch( ( e ) => console.log( e ) )
 		.then( ( result ) => done() ); // Catch PHP errors.
 	});
 
 	it( 'creates new files and directories for objects, components, and modules', ( done ) => {
-		assert.equal( fs.existsSync( config.phpDir + 'objects/o-nav.php' ), true );
-		assert.equal( fs.existsSync( config.phpDir + 'components/c-nav-link.php' ), true );
-		assert.equal( fs.existsSync( config.phpDir + 'modules/breadcrumbs.php' ), true );
+		assert.equal( fs.existsSync( appConfiguration.phpDir + 'objects/o-nav.php' ), true );
+		assert.equal( fs.existsSync( appConfiguration.phpDir + 'components/c-nav-link.php' ), true );
+		assert.equal( fs.existsSync( appConfiguration.phpDir + 'modules/breadcrumbs.php' ), true );
 		done();
 	});
 
 	it( 'parses patterns as expected', ( done ) => {
 		patternShortPaths.forEach( ( shortpath ) => {
 			let expectedContents = fs.readFileSync( expectedDir + shortpath ).toString();
-			let actualContents = fs.readFileSync( config.phpDir + shortpath ).toString();
+			let actualContents = fs.readFileSync( appConfiguration.phpDir + shortpath ).toString();
 
 			assert.equal( expectedContents, actualContents );
 		});
@@ -55,7 +53,7 @@ describe( 'twig to php parser', function() {
 	});
 
 	afterEach( () => {
-		exec( 'rm -r ' + config.phpDir, ( err ) => {
+		exec( 'rm -r ' + appConfiguration.phpDir, ( err ) => {
 			if ( err ) {
 				console.error( err );
 			}